Fixes #68: removed duplicated mcrypt code from Bitauth.php and moved appropriate tests to McryptExtension test
